error catchign

+++ Jenkinsfile
@@ -17,7 +17,9 @@ node {
         //echo pwd()
         dir ('/Users/taylor/ramHacks-2016/'){
           sh "python App/test.py"
-          sh "rm App/spaceGeek/src/*.zip"
+          if(fileExists 'App/spaceGeek/*.zip'){
+            sh "rm App/spaceGeek/src/*.zip"
+          }
           sh "zip -r App/SpaceGeek/master App/spaceGeek/src/*"
         }
       /*   sh "javac App/*.java"
